refactor: move `MastRootHash` and `Interface*` types to their modules;
